requirements.txt
numpy == 1.21.1
pandas == 1.3.0
scipy == 1.7.1
scikit-learn == 1.5.2
implicit == 0.6.1
python-dotenv == 1.0.0
fastapi == 0.115.0
fastapi-utils == 0.7.0
uvicorn == 0.23.2